only organizer can update and delete his event

+++ src/events/events.controller.ts
@@ -2,6 +2,7 @@ import {
   Body,
   Controller,
   Delete,
+  ForbiddenException,
   Get,
   HttpCode,
   Logger,
@@ -21,6 +22,7 @@ import { UpdateEventDto } from './input/update-event.dto'
 import { ListEvents } from './input/list.events'
 import { CurrentUser } from 'src/auth/current-user.decorator'
 import { AuthGuardJwt } from 'src/auth/auth-guard-jwt'
+import { User } from 'src/auth/user.entity'
 
 @Controller('events')
 export class EventsController {
@@ -39,15 +41,15 @@ export class EventsController {
     )
   }
 
-  @Get('practice')
-  practice() {
-    return this.eventsService.practice()
-  }
+  // @Get('practice')
+  // practice() {
+  //   return this.eventsService.practice()
+  // }
 
-  @Get('practice2')
-  async practice2() {
-    return await this.eventsService.practice2()
-  }
+  // @Get('practice2')
+  // async practice2() {
+  //   return await this.eventsService.practice2()
+  // }
 
   @Get(':id')
   async findOne(@Param('id', ParseIntPipe) id) {
@@ -63,26 +65,37 @@ export class EventsController {
   @UseGuards(AuthGuardJwt)
   async create(
     @Body(new ValidationPipe({ groups: ['create'] })) input: CreateEventDto,
-    @CurrentUser() user,
+    @CurrentUser() user: User,
   ) {
     return await this.eventsService.create(input, user)
   }
 
   @Patch(':id')
+  @UseGuards(AuthGuardJwt)
   async update(
     @Param('id') id,
     @Body(new ValidationPipe({ groups: ['update'] })) input: UpdateEventDto,
+    @CurrentUser() user: User,
   ) {
-    return await this.eventsService.update(id, input)
+    return await this.eventsService.update(id, input, user)
   }
 
   @Delete(':id')
+  @UseGuards(AuthGuardJwt)
   @HttpCode(204)
-  async deleteEvent(@Param('id') id) {
-    const result = await this.eventsService.deleteEvent(id)
+  async deleteEvent(@Param('id') id, @CurrentUser() user: User) {
+    const event = await this.eventsService.findOne(id)
 
-    if (result?.affected !== 1) {
+    if (!event) {
       throw new NotFoundException()
     }
+
+    if (user.id !== event.organizerId) {
+      throw new ForbiddenException(
+        null,
+        'You are not authorized to delete this event',
+      )
+    }
+    return await this.eventsService.deleteEvent(id)
   }
 }
